requirements.txt

torch
numpy
matplotlib
requests
python-binance
